requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
# pyTelegramBotAPI==3.6.6
requests==2.21.0
six==1.12.0
urllib3==1.24.2
python-telegram-bot==13.11